Create Dumpassist.sh
Here’s something I created for a client so they don’t call me at 3 AM to debug a cord dumb only to find out. It was nothing but a timing issue.  This Script will read a court down and try and decipher where and when the process went off the rails.